AI_hackathon2024/AI_hackathon2024/nashitora/requirements.txt
Django==4.2
requests
python-dotenv
openai==0.28.0
